+++ CHANGELOG.md
@@ -7,6 +7,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [Unreleased]
 
 ### Added
+- Enable syntax highligting in the query editor (#30)
 - Add command line option to disable logging (#80)
 - Persist column widths across restarts (#53)
 - Add :clearcache command to clear cache files
@@ -14,7 +15,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ## [1.1.0]
 
 ### Added
-- Enable syntax highligting in the query editor (#30)
 - Add query log widget (#32)
 - Add keyboard shortcuts to resize the query editor (#40)